requirements.txt.txt
streamlit
pandas
numpy
joblib
scikit-learn
shap
matplotlib
requests
opencv-python-headless
pyzbar
Pillow
streamlit-webrtc
